BREAKING: changed compiled package structure to allow for multiple exports.
