+++ demo/index.html
@@ -21,6 +21,8 @@
     <h1>Pixi Multistyle Text Demo</h1>
     <main>
       <script>
+        var MultiStyleText = pixiMultistyleText.MultiStyleText;
+
         function appendPixiRendererToId(
           id,
           width = 600,
