+++ package.json
@@ -63,8 +63,8 @@
   },
   "scripts": {
     "demo": "npm run build && opener demo/index.html",
-    "dev": "microbundle watch  --name MultiStyleText --globals 'pixi.js'=PIXI",
-    "build": "microbundle --name MultiStyleText --globals 'pixi.js'=PIXI",
+    "dev": "microbundle watch --globals 'pixi.js'=PIXI",
+    "build": "microbundle --globals 'pixi.js'=PIXI",
     "prepublishOnly": "npm run build",
     "release": "standard-version",
     "test": "jest",
