+++ src/RichText.ts
@@ -1,0 +1,2 @@
+const stub = {};
+export default stub;
